+++ lib/vagrant-omnibus/action/install_chef.rb
@@ -33,13 +33,22 @@ module VagrantPlugins
         def call(env)
           desired_chef_version = env[:global_config].omnibus.chef_version
 
+          if ENV["http_proxy"]
+            http_proxy = "export http_proxy=" + ENV["http_proxy"]
+            https_proxy = "export https_proxy=" + ENV["https_proxy"]
+            if ENV["no_proxy"]
+              no_proxy = "export no_proxy=" + ENV["no_proxy"]
+            end
+            env[:ui].info("proxy use\n#{http_proxy}\n#{https_proxy}\n#{no_proxy}")
+          end
+
           unless desired_chef_version.nil?
             env[:ui].info("Ensuring Chef is installed at requested version of #{desired_chef_version}.")
             if env[:installed_chef_version] == desired_chef_version
               env[:ui].info("Chef #{desired_chef_version} Omnibus package is already installed...skipping installation.")
             else
               env[:ui].info("Chef #{desired_chef_version} Omnibus package is not installed...installing now.")
-              env[:ssh_run_command] = install_chef_command(desired_chef_version)
+              env[:ssh_run_command] = install_chef_command(desired_chef_version,http_proxy,https_proxy,no_proxy)
             end
           end
 
@@ -48,12 +57,15 @@ module VagrantPlugins
 
         private
 
-        def install_chef_command(version=latest)
+        def install_chef_command(version=latest,http_proxy,https_proxy,no_proxy)
           <<-INSTALL_OMNIBUS
+#{http_proxy}
+#{https_proxy}
+#{no_proxy}
 if command -v wget &>/dev/null; then
-  wget -qO- #{INSTALL_SH} | sudo bash -s -- -v #{version}
+  wget -qO- #{INSTALL_SH} | sudo -E bash -s -- -v #{version}
 elif command -v curl &>/dev/null; then
-  curl -L #{INSTALL_SH} -v #{version} | sudo bash
+  curl -L #{INSTALL_SH} -v #{version} | sudo -E bash
 else
   echo "Neither wget nor curl found. Please install one and try again." >&2
   exit 1